+++ spacepackets/ecss/tm.py
@@ -20,6 +20,7 @@ from spacepackets.ccsds.spacepacket import (
     PacketType,
     SpacePacket,
     AbstractSpacePacket,
+    SequenceFlags,
 )
 from spacepackets.ccsds.time import CdsShortTimestamp, read_p_field, CcsdsTimeProvider
 from spacepackets.ecss.conf import (
@@ -228,9 +229,7 @@ class PusTelemetry(AbstractPusTm):
     CDS_SHORT_SIZE = 7
     PUS_TIMESTAMP_SIZE = CDS_SHORT_SIZE
 
-    # TODO: Better timestamp abstraction.
-    #       Use a better default constructor which takes a parameter aggregation
-    #       Supply this constructor as a classmethod in reduced form
+    # TODO: Supply this constructor as a classmethod in reduced form
     def __init__(
         self,
         service: int,
@@ -261,6 +260,7 @@ class PusTelemetry(AbstractPusTm):
             ccsds_version=packet_version,
             data_len=data_length,
             seq_count=seq_count,
+            seq_flags=SequenceFlags.UNSEGMENTED,
         )
         self.pus_tm_sec_header = PusTmSecondaryHeader(
             service=service,