+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - (breaking): `AbstractPusTm` `get_sp_header` renamed to `sp_header` and is a property now.
 - (breaking): `ecss.PusTelemetry`: public member `sp_header` is now named `space_packet_header` to
   avoid name clash with new property.
+- (breaking): `SequenceFlags` argument removed from `ecss.tc.PusTelecommand`. ECSS specifies this
+  field is always set to `SequenceFlags.UNSEGMENTED`.
 
 ## Added